# Conflicts:
#	fe/fe-core/src/main/java/com/starrocks/statistic/StatisticExecutor.java

+++ fe/fe-core/src/main/java/com/starrocks/master/ReportHandler.java
@@ -66,14 +66,12 @@ import com.starrocks.task.CreateReplicaTask;
 import com.starrocks.task.DropReplicaTask;
 import com.starrocks.task.MasterTask;
 import com.starrocks.task.PublishVersionTask;
-import com.starrocks.task.PushTask;
 import com.starrocks.task.StorageMediaMigrationTask;
 import com.starrocks.task.UpdateTabletMetaInfoTask;
 import com.starrocks.thrift.TBackend;
 import com.starrocks.thrift.TDisk;
 import com.starrocks.thrift.TMasterResult;
 import com.starrocks.thrift.TPartitionVersionInfo;
-import com.starrocks.thrift.TPushType;
 import com.starrocks.thrift.TReportRequest;
 import com.starrocks.thrift.TStatus;
 import com.starrocks.thrift.TStatusCode;
@@ -411,10 +409,7 @@ public class ReportHandler extends Daemon {
             // 1. CREATE
             // 2. SYNC DELETE
             // 3. CHECK_CONSISTENCY
-            if (task.getTaskType() == TTaskType.CREATE
-                    || (task.getTaskType() == TTaskType.PUSH && ((PushTask) task).getPushType() == TPushType.DELETE
-                    && ((PushTask) task).isSyncDelete())
-                    || task.getTaskType() == TTaskType.CHECK_CONSISTENCY) {
+            if (task.getTaskType() == TTaskType.CREATE || task.getTaskType() == TTaskType.CHECK_CONSISTENCY) {
                 continue;
             }
 
@@ -731,8 +726,15 @@ public class ReportHandler extends Daemon {
                             continue;
                         }
 
-                        tablet.deleteReplicaByBackendId(backendId);
-                        ++deleteCounter;
+                        // Defer the meta delete to next tablet report, see `Replica.deferReplicaDeleteToNextReport`
+                        // for details.
+                        if (replica.getDeferReplicaDeleteToNextReport()) {
+                            replica.setDeferReplicaDeleteToNextReport(false);
+                            continue;
+                        } else {
+                            tablet.deleteReplicaByBackendId(backendId);
+                            ++deleteCounter;
+                        }
 
                         // remove replica related tasks
                         AgentTaskQueue.removeReplicaRelatedTasks(backendId, tabletId);
@@ -742,9 +744,9 @@ public class ReportHandler extends Daemon {
                                 indexId, tabletId, backendId);
 
                         GlobalStateMgr.getCurrentState().getEditLog().logDeleteReplica(info);
-                        LOG.warn("delete replica[{}] in tablet[{}] from meta. backend[{}], report version: {}"
-                                        + ", current report version: {}",
-                                replica.getId(), tabletId, backendId, backendReportVersion,
+                        LOG.warn("delete replica[{}] with state[{}] in tablet[{}] from meta. backend[{}]," +
+                                        " report version: {}, current report version: {}",
+                                replica.getId(), replica.getState().name(), tabletId, backendId, backendReportVersion,
                                 currentBackendReportVersion);
 
                         // check for clone
